+++ previous-versions/banking/RELEASES.md
@@ -511,3 +511,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.27.1] https://pypi.org/project/codat-banking/0.27.1 - banking
+
+## 2023-09-08 00:02:44
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.82.0 (2.107.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.0] previous-versions/banking
+### Releases
+- [PyPI v0.1.0] https://pypi.org/project/codat-banking/0.1.0 - previous-versions/banking